Store player position and camera position as floats
Fixed jerkiness

Fix a few casting errors

+++ camera.c
@@ -3,7 +3,7 @@
 
 #include "camera0.h"
 
-Camera make_camera(int32_t x, int32_t y) {
+Camera make_camera(float x, float y) {
 	struct Camera *cam = malloc(sizeof(struct Camera));
 	cam->centre_x = x;
 	cam->centre_y = y; 
@@ -15,7 +15,7 @@ void free_camera(Camera cam) {
 	cam = NULL;
 }
 
-void pan_camera(Camera cam, int32_t dx, int32_t dy) {
+void pan_camera(Camera cam, float dx, float dy) {
 	cam->centre_x += dx;
 	cam->centre_y += dy;
 }
